Fix MSVC and GCC warnings unmasked by the previous commit
- fiff_stream.cpp: the HPI coil loop reused 'k', which hides an earlier local
  in read_meas_info (MSVC C4456). Renamed to 'e'.
- ioutils.h: both read_eigen_matrix overloads narrowed a size()/rows() result
  into int without a cast (MSVC C4267). The QList overload had the same
  pattern as the std::vector one, so both are fixed.
- test_utils_sphere.cpp: 2.0f * M_PI is a double and truncates when passed to
  uniform_real_distribution<float> (MSVC C4305). Cast M_PI to float.
- test_utils_kmeans.cpp: assigning a fixed size RowVector3d into a row of a
  dynamically sized MatrixXd makes GCC unable to prove the destination has
  three columns and it reports a false -Warray-bounds inside Eigen. Fill the
  matrix coefficient wise instead.

Verified: clang build 0 warnings, tree wide clang -Wshadow sweep reports 0
sites, 181/181 ctest pass.

+++ src/testframes/test_utils_sphere/test_utils_sphere.cpp
@@ -130,7 +130,7 @@ MatrixX3f TestSphere::generateSpherePoints(const Vector3f& center, float radius,
 {
     MatrixX3f points(nPoints, 3);
     std::mt19937 gen(seed);
-    std::uniform_real_distribution<float> dist_theta(0.0f, 2.0f * M_PI);
+    std::uniform_real_distribution<float> dist_theta(0.0f, 2.0f * static_cast<float>(M_PI));
     std::uniform_real_distribution<float> dist_phi(-1.0f, 1.0f);
 
     for (int i = 0; i < nPoints; ++i) {
